feat(app/highlight): add github dark dimmed theme
